feat: Refactor video download and transcript retrieval
This commit refactors the code for downloading YouTube videos and retrieving their transcripts. The code now uses the pytube library to download videos in mp4 format and stores them in the "downloaded_videos" directory. Additionally, the YouTubeTranscriptApi is used to retrieve the transcript of the video, which is stored in the 'transcript' variable. This refactoring improves the functionality of the code by separating the download and transcript retrieval processes and using more efficient libraries for each task.